requirements.txt
fastapi
uvicorn
pydantic
jinja2
python-multipart
openpyxl
weasyprint
python-dateutil
sqlalchemy
passlib[bcrypt]
python-jose[cryptography]
